Moved all runnable into ./bin directory
In preparation for 2 stage command execution where only one command-name
(brat) needs to be visible.

Links in place where the runnable used to be. Convenient if directory is put
in path which makes tab expansion work. Some scripts are not liked out as
the're going to be abstracted one step further (qry*.sh)