Lowered verbosity, added $OUTDIR to write to
Having high verbosity makes it impossible to see if scripts are doing what
they are supposed to: Any debug printout id drowned in VERBOSE messages
drowned in VERBOSE messages.

BIG FAT NOTE:
Note that redirecting stdout is not enough and that tables are created in
this directory when run.

TBD This need to be dealt with if calibrate.sh is going to be used in
production.